Thanks for your contribution and we appreciate it a lot. The following instructions would make your pull request more healthy and more easily get feedback. If you do not understand some items, don't worry, just make the pull request and seek help from maintainers.


## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->
* Optimize the implementation of lightllm ops.
token attention算子优化前:
![image](https://github.com/DeepLink-org/DIOPI/assets/131418410/7df7436f-a981-40df-a6a6-ca77b3528b8c)
token attention算子优化后：
![image](https://github.com/DeepLink-org/DIOPI/assets/131418410/acbc1d61-3b3f-4b1e-a9ba-786894eac548)
token softmax_reducev算子优化前:
![image](https://github.com/DeepLink-org/DIOPI/assets/131418410/a754cea2-e924-4b37-8ae0-f3a5cc3bea52)
token softmax_reducev算子优化后：
![image](https://github.com/DeepLink-org/DIOPI/assets/131418410/3c164c85-e14d-4b12-a621-93e498e9d821)
同一台机器，同一算子执行十次取稳定下来的运行时间数据，这两个算子相较之前的实现大约提速2倍。
token attention算子: 1.5 ms --> 0.7 ms
token softmax_reducev算子: 1.15ms --> 0.63 ms

## Description
<!--- Describe your changes in detail. -->


## Use cases (Optional)
<!--- If this PR introduces a new feature, it is better to list some use cases here, and update the documentation. -->


## BC-breaking (Optional)
<!--- Does the modification introduce changes that break the backward-compatibility of the downstream repositories? -->
<!--- If so, please describe how it breaks the compatibility and how the downstream projects should modify their code to keep compatibility with this PR. -->


## Checklist
**Before PR**:

- [x] I have read and followed the workflow indicated in the [Contributors.md](https://github.com/DeepLink-org/DIOPI/blob/main/Contributors.md) to create this PR.
- [x] Pre-commit or linting tools indicated in [Contributors.md](https://github.com/DeepLink-org/DIOPI/blob/main/Contributors.md) are used to fix the potential lint issues.
- [x] Bug fixes are covered by unit tests, the case that causes the bug should be added in the unit tests.
- [x] New functionalities are covered by complete unit tests. If not, please add more unit test to ensure the correctness.
- [x] The documentation has been modified accordingly, including docstring or example tutorials.

**After PR**:

- [x] CLA has been signed and all committers have signed the CLA in this PR.